requirements.txt
paho-mqtt>=1.6.0
python-dotenv>=0.19.0
